+++ IdentityCommand/Public/Get-IDTenantConfiguration.ps1
@@ -1,0 +1,21 @@
+# .ExternalHelp IdentityCommand-help.xml
+function Get-IDTenantConfiguration {
+    [CmdletBinding()]
+    param( )
+
+    BEGIN {
+
+        $URI = "$Script:tenant_url/TenantConfig/GetCustomerConfig"
+
+    }#begin
+
+    PROCESS {
+
+        #Send Request
+        Invoke-IDRestMethod -Uri $URI -Method POST
+
+    }#process
+
+    END {}#end
+
+}